Release 25/07/2021 Start Deploy Session

+++ TechShopSolution.Application/Catalog/Coupon/CouponService.cs
@@ -41,19 +41,26 @@ namespace TechShopSolution.Application.Catalog.Coupon
         {
             try
             {
+                
                 var coupon = new TechShopSolution.Data.Entities.Coupon
                 {
                     code = request.code,
                     value = request.value,
                     type = request.type,
                     name = request.name,
-                    max_price = request.max_price,
-                    min_order_value = request.min_order_value,
                     end_at = request.end_at,
                     start_at = request.start_at,
                     isActive = request.isActive,
                     quantity = request.quantity,
                 };
+                if (!string.IsNullOrWhiteSpace(request.max_price))
+                    coupon.max_price = double.Parse(request.max_price);
+                else coupon.max_price = null;
+
+                if (!string.IsNullOrWhiteSpace(request.min_order_value))
+                    coupon.min_order_value = double.Parse(request.min_order_value);
+                else coupon.min_order_value = null;
+
                 _context.Coupons.Add(coupon);
                 await _context.SaveChangesAsync();
                 return new ApiSuccessResult<bool>();
@@ -70,6 +77,8 @@ namespace TechShopSolution.Application.Catalog.Coupon
                 var coupon = await _context.Coupons.FindAsync(id);
                 if (coupon != null)
                 {
+                    if(await _context.Orders.AnyAsync(x=>x.coupon_id == id))
+                        return new ApiErrorResult<bool>($"Bạn chỉ có thể xóa mã giảm giá chưa được ai sử dụng");
                     _context.Coupons.Remove(coupon);
                     await _context.SaveChangesAsync();
                     return new ApiSuccessResult<bool>();
@@ -143,13 +152,19 @@ namespace TechShopSolution.Application.Catalog.Coupon
             };
             return new ApiSuccessResult<CouponViewModel>(couponn);
         }
-        public ApiResult<CouponViewModel> GetByCode(string code)
+        public async Task<ApiResult<CouponViewModel>> UseCoupon(string code, int cus_id)
         {
             var coupon = _context.Coupons.Where(x=> x.code.Equals(code)).FirstOrDefault();
             if (coupon == null)
             {
                 return new ApiErrorResult<CouponViewModel>("Mã giảm giá không tồn tại");
             }
+            var isUse = await _context.Orders.AnyAsync(x => x.cus_id == cus_id && x.coupon_id == coupon.id);
+            if(isUse)
+            {
+                return new ApiErrorResult<CouponViewModel>("Bạn đã sử dụng mã này rồi");
+            }
+
             var couponn = new CouponViewModel()
             {
                 code = coupon.code,
@@ -176,8 +191,15 @@ namespace TechShopSolution.Application.Catalog.Coupon
                     Coupon.name = request.name;
                     Coupon.isActive = request.isActive;
                     Coupon.quantity = request.quantity;
-                    Coupon.max_price = request.max_price;
-                    Coupon.min_order_value = request.min_order_value;
+
+                    if (!string.IsNullOrWhiteSpace(request.max_price))
+                        Coupon.max_price = double.Parse(request.max_price);
+                    else Coupon.max_price = null;
+
+                    if (!string.IsNullOrWhiteSpace(request.min_order_value))
+                        Coupon.min_order_value = double.Parse(request.min_order_value);
+                    else Coupon.min_order_value = null;
+                 
                     Coupon.start_at = request.start_at;
                     Coupon.end_at = request.end_at;
